requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
